requirements.txt
Flask==0.12.2
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
# requires: libldap2-dev, libsasl2-dev
python-ldap==3.0.0
